backend/requirements.txt
# Server
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==1.10.15
python-dotenv==1.0.1
aiofiles==24.1.0

# PDF
pypdf==4.3.1

# LangChain ecosystem
langchain==0.2.12
langchain-community==0.2.11
langchain-google-genai==1.0.7
langchain-text-splitters==0.2.2

# Embeddings (HuggingFace)
sentence-transformers==2.2.2
transformers==4.34.0

# Vector DB
faiss-cpu==1.9.0.post1
numpy==1.26.4

# Utilities
requests==2.32.5
